+++ src/AiToys/App.xaml.cs
@@ -12,20 +12,7 @@ public partial class App : Application
     /// <summary>
     /// Initializes a new instance of the <see cref="App"/> class.
     /// </summary>
-    public App()
-    {
-        InitializeComponent();
-    }
-
-    /// <summary>
-    /// Invoked when the application is launched normally by the end user.  Other entry points.
-    /// </summary>
-    /// <param name="args">Details about the launch request and process. </param>
-    protected override void OnLaunched(Microsoft.UI.Xaml.LaunchActivatedEventArgs args)
-    {
-        var mWindow = new MainWindow();
-        mWindow.Activate();
-    }
+    public App() => InitializeComponent();
 }
 
 #pragma warning restore CA1515