+++ src/Extensions.Hosting.WinUi/WinUiContext.cs
@@ -1,0 +1,19 @@
+using Microsoft.UI.Dispatching;
+using Microsoft.UI.Xaml;
+
+namespace Extensions.Hosting.WinUi;
+
+internal sealed class WinUIContext
+{
+    public Window? MainWindow { get; set; }
+
+    public Application? Application { get; set; }
+
+    public required Type MainWindowType { get; set; }
+
+    public DispatcherQueue? DispatcherQueue { get; set; }
+
+    public bool IsLifetimeLinked { get; set; }
+
+    public bool IsRunning { get; set; }
+}