+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file. See [versionize](https://github.com/versionize/versionize) for commit guidelines.
 
+<a name="0.1.0"></a>
+## [0.1.0](https://www.github.com/ggwozdz90/ai-toys/releases/tag/v0.1.0) (2025-03-07)
+
+### Features
+
+* add Extensions.Hosting.WinUi project and integrate with AiToys ([529ed9d](https://www.github.com/ggwozdz90/ai-toys/commit/529ed9ddb87436206858fd2c2d116d1b619f884d))
+* add Extensions.Hosting.WinUi project and integrate with AiToys ([39509a2](https://www.github.com/ggwozdz90/ai-toys/commit/39509a2dad65b1e8af4635c3606dd443485ea12c))
+* update application startup to use Generic Host for WinUI ([7e0897c](https://www.github.com/ggwozdz90/ai-toys/commit/7e0897ca720da1594c64a94422b2c6069b6c77e9))
+* update application startup to use Generic Host for WinUI ([f270d0e](https://www.github.com/ggwozdz90/ai-toys/commit/f270d0e2755f194ffb38011ed492f26d4615ebce))
+
 <a name="0.0.1"></a>
 ## [0.0.1](https://www.github.com/ggwozdz90/ai-toys/releases/tag/v0.0.1) (2025-03-01)